DataGrid: fix Tab press handler in filter range (T1312521)
